refactor(telegram): split telegram.ts into parse + invoke-turn + wiring
telegram.ts had grown into a 715-line "domain god-file" mixing three
concerns with different change rates and test styles. Split by concern
(pure move, zero behavior change — all 346 tests pass unchanged):

- parse.ts — PURE Telegram→Agent decode: message field extraction, the
  prompt envelope, and the summon/route policy. No state, no IO, no Bot
  API; tested as plain functions. Owns the public TelegramMessage/Update/
  Route types, telegramEnvelope, defaultTelegramRoute.
- invoke-turn.ts — run one turn: assemble inputs (resolve attachments —
  download/vision) and stream agent.invoke.
- telegram.ts — now only ingress + the per-turn lifecycle + composition.

The public subpath surface (@kid7st/fastagent/telegram) is unchanged:
telegram.ts re-exports the moved public symbols, so the shim is untouched.
Companion docs updated (AGENTS.md map, core.md §9.2 module table).

715 → ~440 for telegram.ts; the pure layer is now independently readable
and testable. This straightens structure only — it does not reduce the
essential complexity, which lives in the (unchanged) leaf modules.
