+++ docs/design/core.md
@@ -184,7 +184,9 @@ subsystem owns its invariants in its own module:
 
 | Module | Owns |
 |---|---|
-| `telegram.ts` | the Telegram DOMAIN: ingress (secret token, body cap), summon policy, envelope/prompt assembly, composition |
+| `telegram.ts` | the Telegram WIRING: ingress (secret token, body cap), the per-turn lifecycle (dequeue → ceiling → fold → stream → commit/remove), and composition of the modules below. Pure parsing and turn assembly are split out (next two rows) so the factory holds only stateful wiring |
+| `parse.ts` | PURE Telegram→Agent decode: message field extraction, the prompt envelope, and the summon/route policy. Defining invariant is purity — no state, no IO, no Bot API — tested as plain functions. Owns the public `TelegramMessage`/`Update`/`Route` types, `telegramEnvelope`, `defaultTelegramRoute` (re-exported through `telegram.ts`) |
+| `invoke-turn.ts` | run one turn (the IO half of Telegram→Agent): assemble its inputs — resolve attachments (download files to disk, load vision images; primary throws, buffered degrades per-attachment) — and stream `agent.invoke` with the assembled prompt. `invokeTurn` is the export |
 | `turn-queue.ts` | in-memory per-session serial execution: one turn at a time per session (FIFO chains), different sessions concurrent — the group-UX queue atop the engine lease (a second summon waits instead of colliding as "busy"). Channel-neutral by design (`label` names the consumer). Holds no durability itself; `turn-store.ts` layers it on (see below) |
 | `turn-store.ts` | durable turn intent (L1): persists an accepted turn pre-ACK, removes it when the turn ends (the runner's `finally`), and replays a crash-surviving one on the next start. Recovers the ACKed-but-unfinished window the queue drops; at-least-once, with a replay ceiling for poison turns. Exactly-once / step-replay (L2) is still the K-axis backend |
 | `context-buffer.ts` | un-summoned group discussion: persisted before each webhook ACK (an ACKed update is never redelivered), folded into the next answered turn, cleared only on that turn's `completed` (only then is it provably in the durable session); carries message ids / reply relations / attachment file_ids so later references ("the file Bob sent") resolve |
